+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.2.0] - 2026-08-23
+
 ### Changed
 
 - `mpvd push` starts the daemon if not already started.
@@ -105,7 +107,8 @@ and this project adheres to
 - `remove`/`rm` subcommand
 - `send` subcommand
 
-[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.1.1...main
+[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.0...main
+[2.2.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/rasmusmerzin/mpvd/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/rasmusmerzin/mpvd/compare/v1.0.0...v2.0.0